refactor: integrate `utils.py` into scraper base class
- Removed `utils.py` and its functions
- Create `ScraperBase` class
- Added functions to the base class for all scrapers
- Enabled organized access to static functions
